Minor refactoring
