+++ src/main/java/com/almod/util/PayloadLog.java
@@ -1,12 +1,12 @@
-package com.almod.camel;
+package com.almod.util;
 
+import com.almod.camel.TaskProcessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class PayloadLog {
     private static final Logger LOG = LoggerFactory.getLogger(PayloadLog.class);
@@ -32,8 +32,8 @@ public class PayloadLog {
     private static List<String> getInfoByProcessedTypeDocList() {
         List<String> list = new ArrayList<>();
         list.add("A BATCH OF FILES HAS BEEN GENERATED");
-        for(Object key : localMap.keySet()) {
-            list.add(String.format("Number of %s files: %s", key, localMap.get(key)));
+        for(Object typeOfDoc : localMap.keySet()) {
+            list.add(String.format("Number of %s files: %s", typeOfDoc, localMap.get(typeOfDoc)));
         }
         list.add("Time handling batch files: " + (System.currentTimeMillis() - TaskProcessor.getStartTimeHandlingBatchFiles()) + " ms");
